Split the Open Schema example into two files
