+++ packages/open-schema-type-script/README.md
@@ -13,8 +13,9 @@ An Open Schema implementation that is used to iterate on the Open Schema specifi
 # One time install
 npm ci
 
-# Run the Open Schema example: It will output files that contain engine event information
-npx ts-node packages/open-schema-type-script/src/example/index.ts
+# Run the Open Schema example(s): They will output files that contain engine event information
+npx ts-node packages/open-schema-type-script/src/example/example1.ts
+npx ts-node packages/open-schema-type-script/src/example/example2.ts
 ```
 
 ## Terminology
